Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dhrumilankola <dhrumilankola843@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uagents = "^0.16.2"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
fastapi = "^0.115.2"
fastapi-cors = "^0.0.6"
uvicorn = ">=0.30.1,<0.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
